# Optimize the code