* Remove buf_indices
* Remove dummy node when tmplist is empty

```
jieba cut no hmm        time:   [6.7083 us 6.8836 us 7.0994 us]
  1 Optimize the code
                        change: [-4.0774% -1.9521% +0.6946%] (p = 0.11 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  7 (7.00%) high mild
  5 (5.00%) high severe

jieba cut with hmm      time:   [9.0631 us 9.1854 us 9.3341 us]
                        change: [-5.0586% -2.7487% -0.8670%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) high mild
  7 (7.00%) high severe
```

```
➜  jieba-rs git:(optimize-code) ./target/release/weicheng
4172ms
```